backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
groq==0.11.0
langfuse==2.57.4
arize-phoenix==4.29.0
opentelemetry-sdk==1.27.0
openinference-instrumentation-groq==0.1.4
slowapi==0.1.9
pydantic==2.9.2
aiosqlite==0.20.0
pandas==2.2.3
numpy==2.1.2
